Addon needs JSON representation of Element.
